Dank Memer Command Reminder AIO
Dank Memer is a popular currency bot on Discord where you run commands to gain money and climb up the currency leaderboard. I found it helpful to make this tag to make setting reminders easier to work, redeem (if you are supporting the bot via Patreon) and other stuff you do on the bot.